🐛 fix(studio): reorganize avatars page components and remove deprecated shop page
