+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/AvatarForm/AvatarFormView.tsx
