+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/AvatarForm/index.tsx
