+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/AvatarsTableView.tsx
