+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/DeleteAvatarDialog/DeleteAvatarDialogView.tsx
