+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/index.tsx
