+++ apps/studio/src/ui/shop/widgets/pages/AvatarsPage/AvatarsTable/useAvatarsTable.ts
