--- apps/studio/src/ui/shop/widgets/pages/Shop/index.tsx
@@ -1,5 +1,0 @@
-import { ShopPageView } from './ShopPageView'
-
-export const ShopPage = () => {
-  return <ShopPageView />
-}
